requirements.txt
streamlit 
ultralytics 
Pillow
numpy
opencv-python
